Moved all examples into one module and named it "examples.py".  Added function to read in bins, depths, and error values from csv files
